refactor piracy code into a module, separate from move.
